Introduce bridges for various environments
Filament may end up running inside of a couple different hosts and will
need to adopt different strategies in how to act. This is a first
attempt to at least break out those specializations where possible.

This also marks the first time we can actually load a component inside
of the Lumieres environment.