Arrange filament to accommodate editing packages
I'm planning on having several packages within filament itself, split
along potential webview/iframe lines depending upon how we want to
do this in Lumiere.

Either way the index.html will be the primary filament entry-point
for Lumiere for exploring and editing a package.

The sub-packages, component-editor and package-explorer, will be
for editing an individual component within a package and for exposing
the available components within a package,
respectively.