Reduce number of extraneous packages we have
Filament provides several applications, seeing as each application
relies on assets etc. that are part of the filament package they should
either rely on filament as a dependency or be part of the filament
package.

Seeing as packages don't help us with assets, it's not worth making
these applications packages as they don't stand alone.

I should also remove the package for the component-editor, but that's
a slightly bigger job given how it was originally developed as the
filament app itself.