Detect LibraryItems from project dependencies
This is a first pass that only detects components in the dependencies
of the loaded package. It does not include non-components that could be
added to the template, nor does it consider the loaded package itself.

This also replaces the components.js with sets of LibraryItems provided
in a plugin directory. This step was somewhat of a refactoring of
the existing components.js system, more refinements will come as
the whole plugin system is explored.