Move saving to filament packageEditor
This used to be the responsibility of the componentEditor when that
was all we had. Capturing the save event should probably remain where
I've put it in the packageEditor, if only so it can coordinate more
interesting saving scenarios. The actual saving still feels like it
could end up moving to either the componentEditor or the editingDocument
itself seeing as the editingDocument is so involved in saving.

The low-level "write data to a file" promise is bubbled up to the
packageEditor so we can react to success/failure. This should eventually
be replaced with a higher level promise to save content to a path.