Rely on filament plugin discovery in Lumieres
Now we don't throw a bunch of 404s by checking the filament/plugins
directory, though I still defend that as a quick proof of concept
given how long it takes to do it otherwise…

But the end result of this is quite nice even if the plugin format
and API still needs to get nailed down.

Of course by introducing promises they're now everywhere though-out
the filament loading/opening a project/populating library item APIs.

That's not a bad thing at all.