Enable component creation from package-explorer
Still need to enable menu item support, but this much is working now.
I like the idea of the final committing something to the filesystem
being handled by native components, it feels very official.

We may want to further customize the "creation" of a component with
some interfaces provided prior to presenting the final save panel in
the future.