Promoted project related APIs from main component
I'm not sure what our proposal for well structured montage
applications is, but this was starting to push the limits of what I
was comfortable with being inside a component.

I have promoted project editing lifecycle APIs to a ProjectController.
This controller is not tied to montage components which makes for a
clean slate when it comes to its own lifecycle, and should make for
much easier testing, especially as I have minimized object creation
to only the few areas where we act as a factory.

I also made many improvements to the promises for loading a project
and discovering plugins that should be back-ported to the main
component in the event that I don't follow through with this exact
approach.