Accommodate EditingProxy introduction to Palette
Most striking is the change to how objects are added to the template,
by deserializing in situ.

Otherwise, we just need to make sure we accommodate EditingProxy objects
wherever we expected the object form the stage before. Id like for this
to be seamless, to use the two interchangeably, but I'm not sure it's
worth it. If it's editingProxies all the way down, that's not a problem.