Show components inside the open package in library
Let the recursion begin. (We probably should prevent adding a component
as an instance to itself.)

Either way this works, but won't be save-able without some massaging
as the actual model while editing is "client" from the palette/stage
bootstrapping/package injection. "client" would need to be removed
when saving for runtime, and then…added again for editing?

We need to remove all differences like this between editing and runtime.