Reorg before using substitution to switch docs
The original filament app had been preserved as its own app-in-an-app
as a first step in moving filament over to being a package editor at its
top-level.

Now the component-editor is a single component that provides the
workbench and some controls directly related to the stage. It is this
component that will be placed inside a substitution, one componentEditor
per component we have open.

Switching between open components will simply switch the active 
componentEditor in the substitution. This means that all other editing
panels are not considered part of the componentEditor itself and
will be used to interact with whichever componentEditor is currently
active.