Manually switch between componentEditors
The substitution, for better or worse, actually substitutes one
component for another by removing old content form the DOM. This
may change in the future but is unlikely given how we consider this to
be intrinsic to the substitution, unlike the condition component
that can be configured to remove or hide content.

Either way, we now keep the componentEditors in the DOM all the time.
We still need a way to close one and we should only ever see one at a
time.

Eventually we may want a way to launch filament in a one-component-only
mode to facilitate multiple-window editing.