Make document tabs reflect state correctly
This should work in both directions now, opening new documents will be
reflected in the tabs, and selecting via tabs will correctly show
the expected document.

This also cleaned up the main component a bit more; information
regarding what is open is now centralized and only exposed on the
projectController.