Move find all files in a directory to the server
There was no way to send the guard function along over the q-connection,
and well this is probably faster. Of course it's much faster as we no
longer deeply list the entire node_modules directory.

It's not a recursive walking so I'm not getting the directories, but
we don't quite have a treeController/view just yet anyway.