Always call load on an editor to open a fileUrl
Leave it up to the individual editors to maintain their own internal
set of "open" documents; if they can cache and reuse a document for a
given fileUrl then they need to do that themselves.

This also fixes the standby classes on the documentEditor within the
componentEditor.

I'd like to do something to help encapsulate this document management
so each editor doesn't need to do all of it themselves.