Add rough support for adding non-components
The whole format of "things to add" will change as we end up with
multiple objects being added per serialization. So really the API is
less about adding an object or a component but about adopting some
serialized block.This is enough to get us working with the current
system.