Move canonical openFileUrl to projectController
We no longer wait for the editor to have drawn before we use it to load
the editingDocument; the stage is associated with the editingDocument
when we have both, whenever that happens.

As such, the mainComponent is less involved in the loading process
and while it still needs to insert editors, the responsibility of
determining which editor to use is up to the projectController.

This streamlines the whole opening a bit as we're less dependent on
components rendering to populate the model and controller. It does mean
that we might decide to not reveal any of the editor until it and the
stage have fully rendered, but this is a later task.