Centralize templateObjectsController within editor
Expose it as something that others components can use to not have to
build their own range controller if they don't need to.

In particular the templateExplorer shouldn't be in the business of
exposing a selectedObjects property.

It looks like selection is not working as expected right now, but I'm
removing the "semi-hacks" that are in here in hopes that the
selection should be working correctly now.