Add very rough check for if file is a directory
This is the bit mask checking that node fs does, except I'm doing it
manually with the few bits that I cared about.

We could populate this from node itself, but I had passed along the
stat for better or worse so I figured I'd give this a try and see how
it feels.