Add very rough tree view for file listing
This is extremely rough; if it looks precarious that's because it is.

There are some creative uses of bindings to patch through information
along the template arg/param system into the tree, some code ripped
almost straight from node's FS system for determine if a file is a
directory, and the tree view rendering is suspicious as I'd prefer
nested lists.

I also have no idea how this plays with file watching other than:
it does not.

But it "works" and so it has that in its favor. That and I'm pretty
amazed that once I got it up, I got the listening for expansion and
subsequent population of children working in one go and could
open files etc.

Obviously it's lacking a lot of functionality and need some UI polish.