Update to accommodate Montage UI package schism
This should account for everything. I haven't touched the montage
extension that provides the icons and library items as it's due for an
overhaul and we have some work in that area to deal with the version of
montage the application we've loaded is actually using.

This does mean that we may not be able to add anything to the stage
right now, which wouldn't be surprising given the full breadth and depth
of changes we've just gone through.