Add keyboard shortcut to reveal extension controls
Purely for debugging purposes, the location of this will be elsewhere
as will the mechanics for what you can enable and disable, and how you
go about doing it.

But for testing v0.12 and v0.13 montage editing support I needed this.

control-shift-e to bring up extensions for anybody playing along.