Extract initial startup into appDelegate
This begins to reduce the responsibilities of the projectController;
The applicationDelegate is introduced and begins to take a lead in
staring up the application.

The extensionController is introduced to handle the application-level
task of coordinating loading and activating extensions.

I intend for more code to move over from the projectController into
the applicationDelegate until the projectController could serve as
the prototype of a document controller.

This particular commit is somewhat mid-refactoring.