Extract preview serving from projectController
Now the previewController can focus on this job. There's a lot that's
not working quite right at the moment which could be some fallout from
this reorganization, but the situation is a lot better with the
projectController getting split up.

And with the projectController focused more on managing documents
we'll get to clean that up next.