Introduce documentController to manage documents
For eventual promotion to Montage, this is the very generic document
system that should be able to serve as then backbone for any document
based application.

This also splits the document loading from the component tree.
EditingDocuments now specify an editorType used to edit the document
and provide the mechanism for loading document instances.

Editors are now given documents that have already been loaded.

I have updated everything but the blueprint editor to adopt this new
reality.

There is also plenty of dust that will settle as responsibilities are
more clearly defined for the various components and controllers.