Update to accommodate activeTarget concept
This will need some fleshing out to build the appropriate target chain
but now the mechanism is in place.

The Delete menu item works as an example for what it's worth, but it's
not undoable.

I'm going to now clean up the menu system more thoroughly with the app
working again with the latest montage changes.