Restore some semblance of the undo/redo menu
There's enough wonkiness going on with actually registering the undo
operations that I'm not sure how useful this is.

Namely, simply inspecting an object puts a whole slew of undo operations
onto the stack.

Additionally I still want to look into grouping undo operations as
it's a burden to have compound edits that should be treated as a single
undo operation.