Stop using serialization as data model in proxies
The reelProxy will use the serialization and other specified arguments
as the basis for constructing a model; the serialization is just an
output format.

The model itself is now observable maps and arrays, though we still need
to add quite bit of testing and I'm sure the changes ripple throughout.

This also introduces using mousse and the montage serialization system
to populate the editing model from a montage serialization, and write
a valid montage serialization from the editing model.

It's a nice bit of decoupling that lets us get to work on the binding
explorer.

The LibraryItem concept needs to be revamped to provide a revived
serialization to the proxyObject initialization; adding from the
library and the inspectors need some updating to work with this
fairly drastic change depending upon what internals of the
proxyObject they were relying on.