Begin cleaning up ReelDocument editing API
We're temporarily using addFragment as the API for introducing new
objects to the editingModel. LibraryItems remain as they were, but will
need to become actual templates in the next step to allow including
multiple objects per libraryItem.

Also of note,we now use a simple JSON transferObject for drag and
drop, data not an actual serialization, but that might change. We
couldn't use the actual Montage serialization as it was trying to
find elements in a nonexistent document. We might need to make our
own serializer/deserializer.