Updated reel document editing specs to work
Thankfully not a ton of changes to make here. There's still a lot of
testing to be done for various undoing scenarios in addition to making
`addObjectsFromTemplate` itself undoable.

Also, there are parallel tests that need to be implemented to ensure the
stage-backed representation is kept in sync throughout the various
operations and their undoings.

But we're hopefully at least as good as we were before using
template-based introduction of objects to the editing template.