Move stage manipulation to editingDocument
I want to make the editingController (read: stageController) responsible
for updating the stage with API that seemingly parallels that of the
reelDocument.

Letting the proxyObject edit the stage object didn't jive with that
vision.

This does mean that I'm planning on making the proxy's own
`_setObjectProperty` something that is used by the editingDocument
only.

Along that line of thought the proxy probably doesn't need to know
about its stageObject.