Add support for an exitModalEditor event
To be emitted by the modal editors right now when they decide they
should close.

Filament should probably provide some standard frame with the expected
exit controls.

Right now we listen for escape, but really that's a bit heavy-handed,
escape might be used to cancel an action, not dismiss something as
large as a modal editor.