Mark owner prototype identification as a TODO
This will require some work and probably some UI once we determine
how to detect or allow an author to specify the prototype of the owner.

We do want to know the prototype so we can present an appropriate
blueprint for the owner being edited; typically don't want to write 
the prototype into the declaration though.