Add API, spec, and UI for appending nodes
Some aspects of this are quire rough, in particular the lack of a
proper tree involves some expensive rebuilding as nodes are appended.

I'm getting this in here right now primarily to lock in the changes
to the object removal APIs, which no longer remove a component's
element from the template DOM. This had some effect on the how we
registered undoing removals.

I'll need to keep cleaning this a bit but wrinkles surface best as I
try actually get some of this working so I'm going to get the rest of
the API we need and then clean up, not that I'm making things a
complete mess, but it's a little half baked.

Also, right now the filament UI is hardcoded to only create DIV
elements; we'll need some UI to present a field for specifying the
element, and probably some for changing it after the fact as well.