Add API, specs, UI for insertBefore
I'm hoping this API survives the move to a better UI but I think there
will be some shuffling; I suspect it will be very useful to create node
proxies, and insert them in the tree even without them having an
associated node in the template DOM, much like reelProxies represent
the idea of a templateObject but don't need an actual backing instance 
of that object to be useful.

Anyway we can insert before now.