Add a special anchored representation of the owner
We'd considered this before, and I think it makes sense, unfortunately
it looks pretty awful right now.

Eventually we'll provide a convenient blueprint editing experience for
the owner of the component reel you're editing.

This also gives us a home for introducing mock data into the exposed
properties to see how the rest of the templateObjects react, and
arguably, to see how the stage reacts as well.