Allow dragging onto domExplorer insertion markers
There's a lot of room for improvement here as the actual interaction is
far from perfect, but it's working and the API is there underneath to
at least get us over this hump.

There's still some demo code that needs to go and we'll need to verify
undo and redo operations for all of this.