Stop decoding the URI of the path we open
I initially was about to fix this by encoding the result, but I don't
see any need to decode it in the first place.

This was causing problems where the result was not an encoded URI.
In particular path names with spaces were a problem that caused Lumieres
to crash.