Don't render %20 in recent welcome screen projects
Closes gh-17

Added a converter to the project's name to decodeURI and encodeURI so
that spaces (and accents) would display in a human friendly way in the
project list, even if the projects folders themselves contain %20 etc.

Debugged with project names:
* create project with accent: (château) result: (ch%C3%A2teau)
* created project with space (test 3) result: (test%203)
* ran filament tests 104specs, 0 failures

In need of code review:
* there was already a converter for the url, but decided to try a
separate one for the name to not mess up existing functionality
* the name appears to be only used to display in the projects lists,
but if i modified that directly with decodeURI lumieres crashed, so I
decided to add a converter.