Updating package dependencies
Just to make sure builds are in order right now. We'll need to replace
these with tags/refs when we actually tag a release.

I'm doing this for right now instead of specifying `master` so we have
some ability to go back in time and build an older version of filament
and get the minimum version of the dependencies we want for it to run.