Fix issue with changing ReelProxy's element
NodeProxy was only observing changes in the ReelDocument's list of ReelProxys. This meant that if a ReelProxy had its properties.get("element") changed it would not be picked up by the NodeProxy.

Changing the observer into a binding makes sure that a NodeProxy is always pointing to the right component no matter how the change happened (new ReelProxy in the list or a change in the element).