Assign a data-montage-id to an element
Allows for dropping components on elements without a data-montage-id by automatically creating and assigning one.

Changed the addAndAssignLibraryItemFragment signature from receiving a montageId to a nodeProxy.
Factored out the generic logic of _generateLabel that generates a unique name into a function (_generateUniqueName) so it could be used in the new _generateMontageId function.

Add new setNodeProxyMontageId function that goes into the nodeProxy's element and changes its attributes, it also has undo manager support and dispatches montageId changes in nodeProxy.