Remove vestiges of templateExplorer pan and drag
The real culprit here is the pointer claiming; it was causing problems
when mousing down on the icon, and expecting it to select.

Refusing to surrender the pointer was probably causing problems with
the repetition. To solve this, I've simply removed the panning
and dragging code as well as this mousedown that was intended
to prevent dragging the card when we were trying to drag the reference
icon.

So somewhat overkill, but we'll need to revisit how we actually code
these interactions such that they work.