Improve reacting to filesystem changes
The packageExplorer now does the minimal amount of work necessary
to reflect changes in the underlying fileystem as detected by watchr.

This means we no longer replace the entire tree whenever we detect any
change. This avoids a problem with the tree view which was not
redrawing properly when content was replaced, but it's not a solution
to that problem; just an optimization that happens to avoid the issue
entirely.

I am pretty certain filesInTreeAtUrl API is very rough still, but it
works well enough in testing.