Prevent passing proxies through to live app
This should be enough to get through reported problems, but I'm sure
we'll need a more generic detection of editing objects and generic
API for finding an appropriate replacement.

I imagine we have the same problem with nodeProxies. I'm noting that
as a problem but not bothering to solve it right now.