Make listenerJig appear in overlay near invocation
In addition to appearing near where it was needed, it's more useful
with support for enter to commit and escape to cancel.

There are some issues with the overlay positioning and activeTarget
setting blocking this from really being polished.

Even as it is it's an improvement. I'll bring the bindingsJig into an
overlay as well to make it more uniform before embarking on improving
the event listener experience.