Add experimental drag event from target to listener
This was actually among the first interactions I had in mind when I
put the event button in place with the textfield (to specify/autocomplete
the eventType of the "eventToken"

I really like the mental model of emitting an event FROM one object
and distributing it TO another, but it's actually the exact opposite of
how you approach the setting up listeners in the DOM.

This also Allow dragging an object reference onto event button

This indicates that that object is interested in listening to events
emitted from this target.

This make more sense form how the addEventListener API works I think, but
may be less obvious as an interaction.

Regardless, there's no reason not to have several ways to setup listeners,
sometimes you just happen to be near the listeners hone you start,
and sometimes you're by the emitter.

Actually clicking on the target right now is among the most useless
means seeing as it doesn't pre-populate anything other than the target.