Promote templateCell header to its own component
This is the first in a series of promotions to make my life easier as
I make a specific component for the owner component that will be
more or less the same as the standard templateObjectCells.

It will be up to the cells to choose what information they present
from the various data components I'm creating.

This has the added benefit of cleaning up the drag and drop code, which
starts to signify to me that this is how this should be done.