Remove the move cursor on drag interactions
This just strikes me as something web apps need to do because people
don't expect to drag, let's not out us as a web application in ways 
we don't need to.

Maybe I just don't like how giant the move cursor is, and I like to
reserve it for actually moving things in space.