Restore same bindingInfo object on removal undo
Prior to this we called the `defineBinding` API with the arguments to
define a new binding that looked for all intents and purposes was the
same binding.

This caused problems downstream within the undoManager where undo
operations might have been registered to work on the original
binding object that had now been removed.

Now the references throughout the undoOperations work as the original
bindingInfo object is never destroy and replaced by a doppelgänger.

I also preserve the index of the binding to prevent hopping around
when undoing, but there may be a value in keeping the bindings organized
alphabetically by targetPath or something instead, eventually.