Scroll newly created objects into view
This works around a bit of a problem where the repetition doesn't help
us know when the iteration has been placed in the right spot and
reflects the object that it's actually representing.

There is a period prior to this where the object is associated with the
iteration as reported, but nothing has happened in the DOM yet. If we
draw at that time, we scroll to where the iteration is currently, not
where it ends up when all is said and done.