Update integration test runner to run all tests
Now, when properly configured, Lumieres will be launched and quit
before and after each test. It's completely hands-free and doesn't stop
on a failing test.

Reporting of failures could be vastly improved as could the exit code
of the runner script itself.

Eventually we should find all the tests worth running within the
directory without instruction.