Allow hiding templateObjects
There's no a control to set the editorMetadata isHidden property on each
template object that is not the owner.

There is also a control for showing or hiding these hidden
templateObjects. All of this UI should be considered temporary.

This ushers in the ability to expose all templateObjects when needed
but otherwise, to hide some by default such as the actionEventListener
which by itself usually adds nothing but clutter and can be represented
in a more useful manner.

In some cases we may even create these objects implicitly, hiding them
by default in some circumstances.