Add option for specifying listener methodName
This optional parameter will implicitly create an actionEventListener
to map the eventHandling through to call the specified method on the
specified listener.

This also creates a quick little reference for the default event handler
methods that would be used in the absence of a directed event handling
method. Eventually we'll be able to detect the presence of such these
methods. Beyond that we'll also be able to put one you select in place.

For now you can copy and paste the name. The UX will need some obvious
improvements.

Now to actually go and create that eventListener…