Create implicit actionEventListener when needed
Now when adding a listener and specifying a custom method to call on
that listener, we'll implicitly create an actionEventListener to handle
the event and call the method on the "listener" specified.

For the most part this means adding an event listener is much more
straight forward regardless of whether you want to use the default
handler method conventions or a directed event handler method.