Guard against numerous un-highlightable components
In particular we were missing access to the editingController
if the highlighting started before the stage was fully initialized.

Additionally, we were stumbling over situations where we tried to
highlight components that had no representation on the stage due to
their ancestry i.e. inside a repetition of a substitution etc.