Add specs to validate undo ability of event APIs
This is not exhaustive but should ensure that in typical situations
the undoing of event API calls that involve an actionEventListener
should behave as expected.

In working on this I have come to realize that the undo blocks don't
do anything to guard against multiple async operations opening up
sunblocks and unintentionally capturing undo registrations from
intended for other undo blocks opened by other async processes.

So that's a problem to look into later.