Improve integration tests with setup and teardown
In addition for me to have a place to setup and teardown tests, this
also streamlines a bit of the error reporting.

I haven't actually implemented any of the setup or tear down though,
they'll probably be node or shell scripts that will be executed prior
to and after running a test to do things like cloning a particular
repository and removing it and other temporary files afterwards.