Add a side-effect free undo of restoring listeners
The `_addOwnedObjectEventListener` was relying on
`removeOwnedObjectEventListener` as its inverse operation; the latter
method has a side-effect of removing any associated AEL.

The end result was that this ended up removing the related AEL, and then
it tried to remove it again as we'd captured that specific removal
as part of the undo batch already. This ended up failing as we can't
remove an object twice.