Refresh stage upon listener edits
This works around a bit of an issue where undoing the addition of an
AEL listener (removing the AEL, removing the listener entry) would
trigger a stage refresh upon removing the AEL (triggering an error as
the listener entry refers to the now missing AEL) and then never
triggered another stage refresh to correct the issue.

This is a temporary fix to work around that issue.