Close remove AEL undo batch as quickly as possible
In this particular scenario there's no need to wait for any promises
to resolve before closing the batch of undos.

In general, the undo registrations themselves are done synchronously,
the resolutions on how to perform the undo are potential asynchronous;
as long as operations that will register undo operations are run
synchronously, their undo operations will be registered correctly.

It's only the other scenarios where we run into problems where those
operations can't begin until some other asynchronous operation has
completed. Then we'd potentially have a batch open while waiting
for that asynchronous operation to complete which is undesired.