Rely on environment bridge provided by host
This removes the browser bridge as a fall back for right now.
Eventually, lumieres should be updated to do the same thing so we can
remove the idea of lumieres from filament completely.

We will need a fallback eventually to handle running filament without
a server, assuming we figure out what that even means or how that would
work.