Correct module name resolving to load document
It looks like the code that was here previously was intended to find a
moduleId and ensure it had no trailing slash.

It was not doing this correctly and was instead passing along a url
as a moduleId, which worked for the most part.

There are still downstream changes that will need to be made as there
is a later call in the editing Frame that is raising a warning as we 
now pass along this valid moduleId as a location to `_getRequireForModuleLocation`