Put project initialization in place for bare repos
The delegate simply puts things in a state where the user needs to
interact in some way to resolve a promise.

We'll certainly need to add more UI around this other specific flows.

We might also need to be more active about the presentation of such
a decision the user needs to make, but I don't think we always need to
be if the rest of the UI is pretty reactive.

e.g. A prompt knows to come forward when the application is in a certain
state and reports some decision made via an event.