Move firefly specific recent project concept
For now at least, the repository linkage/concept is firefly specific.
Eventually, more of this connection to a repository on github, folder
on dropbox, will be exposed to filament proper but I'm trying to avoid
relying on our rushed implementation of these sorts of things until
we've actually thought about them and decided what they provide and
where they live.