Accommodate CodeMirror history management
CodeMirror manages its own undo/redo history, thankfully, but we need it
reflected as undo and redo operations in our own UndoManager to play
nicely with the rest of the environment.

For example, CodeMirror handles cmd-z by default but doesn't react to
the undo performed by a menuItem. While simply calling the `undo` method
of the CodeMirror document would work fine, this would then be a special
case for the menus in our world. Specifically they would need to know
that they're interacting with a CodeMirror document and vary how they
behave in how they check to see if an undo can be performed and how
they actually perform an undo.

It's easier to keep the undo/redo checking and execution consistent
at the higher level by registering operations to our own UndoManager
as they are registered in CodeMirror's history.

There is probably some room for optimizations as the undo and redo
operation promises simply call the CodeMirror Document undo and redo;
we probably don't need to make a brand new promise each time to save
one memory.

Anyway, this should be enough to get working. Most of the heavy lifting
has moved to the CodeEditorDocument to parallel the undo/redo API that
is expected on Document in our world.

I'm sure there's some work to be done as we'll probably approach the
undo and redo stack maximums much quicker in a text editor than we would
elsewhere.