Use labels that match their data-montage-id
When adding a library item to the stage we need to make sure that the
labels of the library item will match the corresponding data-montage-id.
This is divided into two steps: first we merge the DOM node and solve any
conflict that might exist with clashing data-montage-ids. We make sure the
 generated data-montage-ids do not conflict with the labels of the
 serialization.
Second, when merging the serializations, we rename the labels to match the
data-montage-ids that were already merged into the document.