Make library items represent actual templates
Right now a template is generated from bits and pieces that would
normally comprise a valid template. If a library item happens to have
an actual template available we could skip the template
generation step; this change aims to make that happen.

This also diminishes the role a LibraryItem has in describing a specific
single module. It should be assumed that a single module in a package
might have multiple libraryItems that could conceptually represent said
module alone, but probably does so in different configurations.

Furthermore, multiple libraryItems may also include that module as part
of its payload.

In short, the LibraryItems are a transport medium for modules.
Information such as the name and icon used by the libraryItem pertain
only to the library item, and not necessarily the modules included
in the libraryItem's template.

With this change in place libraryItems may become the backbone for
copy/paste, extracting favorite snippets, and refactoring templates.

This change begins maturing the extension and libraryItem
system a bit from its initial implementation. There is still work
to be done.