Pass all template content through for insertion
Prior to this we were expecting to have snippets of parts of a template.
We now have an entire HTML5 template worth of content in a string to
pass along for insertion into the template being edited.

While we still need to reconstitute a template object using said
templateContent, the process opens the doors to a much more streamlined
workflow.

Of note, since a template can be represented as text/plain content it
should be noted that we can now drag in a valid template from any
other application and accept it.

The work here only accommodates dropping from the LibraryCell (and
other text/plain sources) onto the templateExplorer.

Some work should be done to coerce less than perfect templateContent
from other sources into something useful.