Allow dropping templateContent onto domExplorer
Some interesting decisions to make here regarding dropping libraryItems
onto a dom node in what was "assignment" before; we should really
look at the incoming template to see if its content is compatible with
the operation before performing it.

We'll need to discard either the existing node or the incoming
content.