Restrict assigning libraryItems to existing nodes
We only want to be able to do this for libraryItems that have a single
moduleId.

We might want to eventually do further sanity checks based upon the
content itself, but in general if there's a situation that's willing
to accept a serialization and figure out what to do about any relevant
markup, we now have a mime type to represent that content.