Ensure templateObject constructors come package
This is a change needed after sandboxing the require used to find
the templateObject's constructor; we need to use that same require
to load the constructor for the Component, which we use to alter the
display of the templateObjectCell.

This also cleans up invalid exports object avoiding a problem where old
versions of MR would incorrectly return an invalid exports on some
occasions. Fixed described here:
https://github.com/montagejs/mr/pull/66

This avoids the problem if the packaged montage does not have the fix
applied.