Preserve unknown serialization units from original
Basically making use of the new originalSerializationMap now
available on the editingProxy after updating Palette.

This specifically addresses us inadvertently discarding the
localization units. And while those will eventually be something we
promote and offer editing assitance for in the tool, much as we do for
bindings and listeners, we generally shouldn't discard things we don't
understand.