Ignore description if object to describe changed
Prior to this there were some race conditions where we might set
the object to be Foo, go off to describe it, and then set the object to
be Bar, and then go off to describe it.

Of course, the descriptions might resolve in any order; discard
the description for object Foo, or general, whichever one we're not
representing right now.