Don't fail when requested to close same file twice
While it generally shouldn't happen, there's no reason for errors
downstream if we end up being requested to close a file we're in the
process of closing.

Now we simply return the same promise for the file we're closing.

This was happening due to multiple listeners observing the close
key command in the Lumieres environment. I'll have a fix for that but
didn't want to lose the chance to fix a good bug worth safeguarding
against.